refactor(web): move public docs to apps workspace
