+++ apps/web/app/components/skill-catalog.tsx
