+++ apps/web/app/skills/[skillId]/page.tsx
