+++ apps/web/package.json
@@ -1,0 +1,25 @@
+{
+	"name": "hr-skills-web",
+	"version": "0.0.0",
+	"private": true,
+	"scripts": {
+		"docs-data": "bun ../../packages/hr-skills-build/src/build/generate-docs-data.ts",
+		"dev": "bun run docs-data && next dev",
+		"build": "bun run docs-data && next build",
+		"typecheck": "bun run docs-data && tsc --noEmit"
+	},
+	"dependencies": {
+		"hr-skills-build": "workspace:*",
+		"marked": "^18.0.9",
+		"next": "^15.5.0",
+		"react": "^19.2.0",
+		"react-dom": "^19.2.0",
+		"sanitize-html": "^2.17.7"
+	},
+	"devDependencies": {
+		"@types/node": "^20.17.6",
+		"@types/react": "^19.2.0",
+		"@types/react-dom": "^19.2.0",
+		"@types/sanitize-html": "^2.16.1"
+	}
+}
