+++ package.json
@@ -30,6 +30,7 @@
 	"workspaces": {
 		"packages": [
 			".",
+			"apps/*",
 			"packages/*",
 			"playground/*"
 		],
