+++ playground/next-app/package.json
@@ -3,23 +3,18 @@
 	"version": "0.0.0",
 	"private": true,
 	"scripts": {
-		"docs-data": "bun ../../packages/hr-skills-build/src/build/generate-docs-data.ts",
-		"dev": "bun run docs-data && next dev",
-		"build": "bun run docs-data && next build",
-		"typecheck": "bun run docs-data && tsc --noEmit"
+		"dev": "next dev",
+		"build": "next build"
 	},
 	"dependencies": {
 		"hr-skills-build": "workspace:*",
-		"marked": "^18.0.9",
 		"next": "^15.5.0",
 		"react": "^19.2.0",
-		"react-dom": "^19.2.0",
-		"sanitize-html": "^2.17.7"
+		"react-dom": "^19.2.0"
 	},
 	"devDependencies": {
 		"@types/node": "^20.17.6",
 		"@types/react": "^19.2.0",
-		"@types/react-dom": "^19.2.0",
-		"@types/sanitize-html": "^2.16.1"
+		"@types/react-dom": "^19.2.0"
 	}
 }
